% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statfunction.R
\name{lm_function}
\alias{lm_function}
\title{Summary of linear model}
\usage{
lm_function(df, dependent, independent_cols)
}
\arguments{
\item{dependent}{Dependent Variable}

\item{independent_cols}{Independent Variable}

\item{data}{Data frame}
}
\value{
Linear Model summary
}
\description{
Creates a summary of a linear model that can be used to compare variances.
}
