% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutation}
\alias{mutation}
\title{Conversion from grams to kilograms}
\usage{
mutation(data, col, col2)
}
\arguments{
\item{data}{Data frame}

\item{col}{column 1}

\item{col2}{Column 2}
}
\value{
Data frame with a new kilogram weight column
}
\description{
Creates a new column that that has a new weight that has gone through a conversion from grams to kilograms
}
