% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunction.R
\name{plot_graph}
\alias{plot_graph}
\title{Creates a scatter plot}
\usage{
plot_graph(data, x, y)
}
\arguments{
\item{data}{Data frame}

\item{x}{X-axis variable}

\item{y}{Y-axis variable}
}
\value{
Scatter plot of x and y
}
\description{
Creates a scatter plot to compare the linear relationship of two variables.
}
