% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3mfunction.R
\name{tm}
\alias{tm}
\title{Calculates min, max, and mean}
\usage{
tm(data, col1, col2)
}
\arguments{
\item{data}{Data frame}

\item{col1}{Column 1}

\item{col2}{Column 2}
}
\value{
mean, minimum, maximum
}
\description{
Calculates the minumum number, maximum number, and average of a column
}
